Fixed ATA Driver
Fixed an error in the ATA Driver, added reading and writing arbetrary amounts of data to the drive.

+++ kernel/lib/hwio.c
@@ -86,7 +86,7 @@ void p_serial_writebytes(char* bytes, unsigned int size)
 	}
 }
 
-void p_serial_writenum(int num)
+void p_serial_writenum(long num)
 {
 	if(num < 0)
 	{
